Create README.md
This project is a website that allows users to create their own receipts. Users create an account, then generate the receipt they want. The website includes a database containing the format of the receipt. Users enter details such as the name of their company or store, the items or services provided, and the final cost. The interface also displays the username of the individual, with each user having their own unique name. Proper registration is required for users to create their accounts.